app/envio_tienda/requirements.txt
psycopg2-binary
faker
kafka-python-ng
pymongo
prometheus-client
